get stats data to dashboard

+++ app/Http/Controllers/DemandController.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use App\Models\Demand;
 use Illuminate\Http\Request;
 
@@ -17,16 +16,22 @@ class DemandController extends Controller
         $request->validate([
             'fullName' => 'required',
             'message' => 'required',
+            'subject' => 'required',
             'email' => 'required|email'
         ]);
         $demand = new Demand();
         $demand->fullName = $request->fullName;
         $demand->email = $request->email;
+        $demand->subject = $request->subject;
         $demand->message = $request->message;
         $demand->save();
     }
-
-    public function destroy(Demand $demand){
+    public function destroy($id){
+        $demand = Demand::find($id);
+        if (!$demand) {
+            return response()->json(['message' => 'Ocannot delete undefined demand!'], 404);
+        }
         $demand->delete();
+        return response()->json(['message' => 'Demand de stage supprimée avec succès']);
     }
 }
